Fixed rising glitch.
Created CharacterStats class, which stores health, maxSpeed, DamageInfo,
hitboxes, and sends the damage info of the attacker to the victim, and

Create DetectTrigger class, which detects interactions between hitboxes
(colliders) and calls the victim's Injure function.

+++ Assets/Scripts/DamageInfo.cs
@@ -4,9 +4,9 @@ using UnityEngine;
 
 public class DamageInfo {
 
-	public float damageModifier;
+	public int damageModifier;
 
-	public DamageInfo(float damageModifier)
+	public DamageInfo(int damageModifier)
 	{
 		this.damageModifier = damageModifier;
 	}
